+++ CHANGELOG.md
@@ -2,7 +2,9 @@ v0.5.0
 <hr>
 * Added `FileSessionHandler`
 * Added `ISession::getId()` and `ISession::setId()`
-* Removed `User` and `Credentials` from `ISession`
+* Removed `User` and `Credentials` from `ISession` and 'Session'
+* Added `CompiledRoute::getPathVariable()`
+* Fixed type-hint for `Router::getMatchedRoute()` to return type `CompiledRoute`
 * Now uses fully-qualified class names in `use` statements and PHPDoc
 * Changed class names to make them readable without having to be fully-qualified:
   * `RDev\Authentication\Credentials\Credentials` renamed to `RDev\Authentication\Credentials\CredentialCollection`